chore: clean up, fix clippy warnings, update docs
- Fix clippy warnings (auto-fix + manual parens/doc comments)
- Update TRY.md with 3-cycle results (358→366, +11 total)
- Update REPORT.md §6: calling-layer multi-perspective findings
- Improve long_decode.rs doc comments for accuracy
- Remove redundant code patterns across codebase (auto-fix)

Test: 20/20 baseline continuously passing, 0 new warnings

+++ src/main.rs
@@ -62,7 +62,7 @@ fn run_decode(file: &PathBuf, mode: &str, low: f64, high: f64, depth: usize, max
         wav.samples.len() as f64 / wav.sample_rate as f64
     );
 
-    let samples_f32: Vec<f32> = wav.samples.iter().map(|&x| x as f32).collect();
+    let samples_f32: Vec<f32> = wav.samples.iter().map(|&x| x).collect();
     let start = Instant::now();
 
     if mode == "ft4" {
